Refactoring RBAC security scripts

+++ logging/bin/security_create_rbacNEW.sh
@@ -4,8 +4,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 #
-# Creates the following RBAC structures (NST=NAMESPACE|NAMESPACE-__TENANT__):
-# 
+# Creates the following RBAC structures (NST=NAMESPACE|NAMESPACE_TENANT):
+#
 #
 #                                    /-- [ROLE: kibana_user]       (allows access to Kibana)
 # [BACKEND_ROLE: {NST}_kibana_user]<-
@@ -25,6 +25,10 @@
 cd "$(dirname $BASH_SOURCE)/../.."
 source logging/bin/common.sh
 this_script=`basename "$0"`
+
+source logging/bin/rbac-include.sh
+source logging/bin/apiaccess-include.sh
+
 
 function show_usage {
   log_message  "Usage: $this_script NAMESPACE [TENANT]"
@@ -74,7 +78,7 @@ if [[ "$TENANT" =~ -H|--HELP|-h|--help ]]; then
    show_usage
    exit
 elif [ -n "$TENANT" ]; then
-   NST="${NAMESPACE}-__${TENANT}__"
+   NST="${NAMESPACE}_${TENANT}"
 else
    NST="$NAMESPACE"
 fi
@@ -87,12 +91,9 @@ fi
 
 
 INDEX_PREFIX=viya_logs
-#ROLENAME=search_index_$NAMESPACE
-#BE_ROLENAME=${NAMESPACE}_kibana_users
 ROLENAME=search_index_$NST
 BE_ROLENAME=${NST}_kibana_users
 if [ "$READONLY" == "true" ]; then
-   #RO_BE_ROLENAME=${NAMESPACE}_kibana_ro_users
    RO_BE_ROLENAME=${NST}_kibana_ro_users
 else
    RO_BE_ROLENAME="null"
@@ -106,7 +107,6 @@ cp logging/es/odfe/rbac $TMP_DIR -r
 
 # Replace PLACEHOLDERS
 sed -i'.bak' "s/xxIDXPREFIXxx/$INDEX_PREFIX/g"  $TMP_DIR/rbac/*.json                  # IDXPREFIX
-#sed -i'.bak' "s/xxNAMESPACExx/$NAMESPACE/g"     $TMP_DIR/rbac/*.json                  # NAMESPACE
 sed -i'.bak' "s/xxNAMESPACExx/$NST/g"     $TMP_DIR/rbac/*.json                  # NAMESPACE
 
 
@@ -119,164 +119,18 @@ export ES_ADMIN_PASSWD=$(kubectl -n $LOG_NS get secret internal-user-admin -o=js
 tmpfile=$TMP_DIR/output.txt
 
 
-# TO DO: Replace w/function call that uses existing ES connection or establishes one via port-forwarding
+sec_api_url=""
+get_sec_api_url
 
-# set up temporary port forwarding to allow curl access
-ES_PORT=$(kubectl -n $LOG_NS get service v4m-es-client-service -o=jsonpath='{.spec.ports[?(@.name=="http")].port}')
-
-# command is sent to run in background
-kubectl -n $LOG_NS port-forward --address localhost svc/v4m-es-client-service :$ES_PORT > $tmpfile  &
-
-# get PID to allow us to kill process later
-pfPID=$!
-log_debug "pfPID: $pfPID"
-
-# pause to allow port-forwarding messages to appear
-sleep 5s
-
-# determine which port port-forwarding is using
-pfRegex='Forwarding from .+:([0-9]+)'
-myline=$(head -n1  $tmpfile)
-
-if [[ $myline =~ $pfRegex ]]; then
-   TEMP_PORT="${BASH_REMATCH[1]}";
-   log_debug "TEMP_PORT=${TEMP_PORT}"
-else
-   set +e
-   log_error "Unable to obtain or identify the temporary port used for port-forwarding; exiting script.";
-   kill -9 $pfPID
-   rm -f  $tmpfile
-   exit 18
+if [ -z "$sec_api_url" ]; then
+   log_error "Unable to determine URL to access security API endpoint"
+   exit 1
 fi
 
 
-
-function role_exists {
-   #Check if $role role exists
-   #
-   # Returns: 0 - Role exists
-   #          1 - Role does not exist
-
-   local role
-
-   role=$1
-
-   response=$(curl -s -o /dev/null -w "%{http_code}" -XGET "https://localhost:$TEMP_PORT/_opendistro/_security/api/roles/$role"   --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
-   if [[ $response == 2* ]]; then
-      log_debug "Confirmed [$role] exists."
-      return 0
-   else
-      log_debug "Role [$role] does not exist."
-      return 1
-   fi
-}
-
-function create_role {
-   # Creates role using provided role template
-   #
-   # Returns: 0 - Role created
-   #          1 - Role NOT created
-
-   local role role_template
-
-   role=$1
-   role_template=$2
-
-   response=$(curl -s -o /dev/null -w "%{http_code}" -XPUT "https://localhost:$TEMP_PORT/_opendistro/_security/api/roles/$role"  -H 'Content-Type: application/json' -d @${role_template}  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
-
-   if [[ $response == 2* ]]; then
-      log_info "Security role [$role] created [$response]"
-      return  0
-   else
-      log_error "There was an issue creating the security role [$role] [$response]"
-      log_debug "template contents: /n $(cat $role_template)"
-      return 1
-   fi
-}
-
-function ensure_role_exists {
-   # Ensures specified role exists; creating it if necessary
-   #
-   # Returns: 0 - Role exists or was created
-   #          1 - Role does NOT exist and/or was NOT created
-
-   local role role_template
-
-   role=$1
-   role_template=${2:-null}
-
-   if  role_exists $role; then
-      return 0
-   else
-      if [ -n "$role_template" ]; then
-         rc=$()
-         if create_role $role $role_template; then
-            return 0
-         else
-            return 1
-         fi
-      else
-         # couldn't create it b/c we didn't have a template
-         log_debug "No role template provided; did not attempt to create role [$role]"
-         return 1
-      fi
-  fi
-
- }
-
-
-function add_rolemapping {
- # adds $ROLENAME and $RO_BE_ROLENAME to the
- # rolemappings for $targetrole (create $targetrole if it does NOT exists)
-
- targetrole=$1
- berole=$2
- log_debug "Parms passed to add_rolemapping function  targetrole=$targetrole  berole=$berole"
-
-
- # get existing rolemappings for $targetrole
- response=$(curl -s -o $TMP_DIR/rolemapping.json -w "%{http_code}" -XGET "https://localhost:$TEMP_PORT/_opendistro/_security/api/rolesmapping/$targetrole"  --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
-
- if [[ $response == 404 ]]; then
-    log_debug "Rolemappings for [$targetrole] do not exist; creating rolemappings. [$response]"
-
-    json='{"backend_roles" : ["'"$berole"'"]}'
-    verb=PUT
-
- elif [[ $response == 2* ]]; then
-    log_debug "Existing rolemappings for [$targetrole] obtained. [$response]"
-    log_debug "$(cat $TMP_DIR/rolemapping.json)"
-
-    if [ "$(grep $berole  $TMP_DIR/rolemapping.json)" ]; then
-       log_debug "A rolemapping between [$targetrole] and  back-end role [$berole] already appears to exist; leaving as-is."
-       return
-    else
-       json='[{"op": "add","path": "/backend_roles/-","value":"'"$berole"'"}]'
-       verb=PATCH
-    fi
-
- else
-     log_error "There was an issue getting the existing rolemappings for [$targetrole]. [$response]"
-     kill -9 $pfPID
-     exit 17
- fi
-
- log_debug "JSON data passed to curl [$verb]: $json"
-
- response=$(curl -s -o /dev/null -w "%{http_code}" -X${verb} "https://localhost:$TEMP_PORT/_opendistro/_security/api/rolesmapping/$targetrole"  -H 'Content-Type: application/json' -d "$json" --user $ES_ADMIN_USER:$ES_ADMIN_PASSWD --insecure)
- if [[ $response != 2* ]]; then
-    log_error "There was an issue creating the rolemapping between [$targetrole] and backend-role(s) ["$berole"]. [$response]"
-    kill -9 $pfPID
-    exit 22
- else
-    log_info "Security rolemapping created between [$targetrole] and backend-role(s) ["$berole"]. [$response]"
- fi
-
-}
-
 #index user
 ensure_role_exists $ROLENAME $TMP_DIR/rbac/index_role.json
-add_rolemapping $ROLENAME $BE_ROLENAME 
+add_rolemapping $ROLENAME $BE_ROLENAME
 
 #kibana_user
 add_rolemapping kibana_user $BE_ROLENAME null
@@ -299,8 +153,10 @@ if [ "$READONLY" == "true" ]; then
 fi
 
 # terminate port-forwarding and remove tmpfile
-log_info "You may see a message below about a process being killed; it is expected and can be ignored."
-kill  -9 $pfPID
+if [ -n "$pfPID" ]; then
+   log_info "You may see a message below about a process being killed; it is expected and can be ignored."
+   kill  -9 $pfPID
+fi
 rm -f $tmpfile
 
 #pause to allow port-forward kill message to appear
@@ -308,13 +164,23 @@ sleep 7s
 
 log_notice "Access controls created [$(date)]"
 echo ""
-log_notice "============================================================================================================================"
+
+log_notice    "================================================================================="
+log_notice    "   Assign users the back-end role of  [${BE_ROLENAME}] to                        "
+log_notice    "   grant them access to Kibana and limit their access to log messages            "
 if [ -n "$TENANT" ]; then
-   log_notice "   Assign users the back-end role of [${BE_ROLENAME}] to grant access to Kibana and limit access to log messages for the [$TENANT] tenant within the [$NAMESPACE] namespace"
+   log_notice "   from the [$TENANT] tenant within the [$NAMESPACE] namespace          "
 else
-   log_notice "   Assign users the back-end role of [${BE_ROLENAME}] to grant access to Kibana and limit access to log messages for [$NAMESPACE] namespace "
+   log_notice "   from the [$NAMESPACE] namespace.                               "
 fi
 if [ "$READONLY" == "true" ]; then
-   log_notice " Assign users the back-end role of [${RO_BE_ROLENAME}] to grant access to log messages for [$TENANT/$NAMESPACE] namespace but limit Kibana access to READ-ONLY ="
+   log_notice "                                                                                 "
+   log_notice "   Assign users the back-end role of  [${RO_BE_ROLENAME}] to                     "
+   log_notice "   grant them READ-ONLY access to Kibana and limit their access to log messages  "
+   if [ -n "$TENANT" ]; then
+      log_notice "   from the [$TENANT] tenant within the [$NAMESPACE] namespace       "
+   else
+      log_notice "   from the [$NAMESPACE] namespace.                            "
+   fi
 fi
-log_notice "============================================================================================================================"
+log_notice    "================================================================================="
